+++ tests/Provider/QueryBuilderProvider.php
@@ -444,6 +444,10 @@ final class QueryBuilderProvider extends \Yiisoft\Db\Tests\Provider\QueryBuilder
         $values['unsigned()'][0] = 'integer';
         $values['scale(2)'][0] = 'numeric(10,2)';
         $values['integer(8)->scale(2)'][0] = 'integer';
+        $values["collation('collation_name')"] = [
+            'varchar(255) COLLATE "C"',
+            ColumnBuilder::string()->collation('C'),
+        ];
 
         $db = self::getDb();
         $serverVersion = self::getDb()->getServerInfo()->getVersion();